Added automatic JSON serialization and deserialization for Tasks
It might not be the greatest idea to require JSON and ONLY JSON, but
it's the fastest serialization format within python (faster than
pickling even), and it's also readable (usually! :v)

On the TODO now is to get the documentation finished for core, and to
update all the unit tests. I'm also looking out for continuous
integration that works with tox, runs on Python 3, and doesn't have a
bad interface (or confusing documentation) like buildbot :/