Added travis-ci configuration file
Added FileList object to utility. This will be used by tasks so that
messing up input and output won't cause an issue for references. A small
test has also been added to the test_utility module, and will be
expanded shortly.

Travis CI is now being used for bit, as such, as a pip requirements file
has been added. It does not currently include:

 * nose
 * distribute
 * virtualenv

as these are provided by the requirements.txt file.