Added several new files that will be modified soon.
Decided to make Mako a dependency for the CXX module. Specifically for
the configure task. I really like how its built, and while it was
written for the web, it lends itself well to general development. It's
basically a mini pre-processor :P.

As such, the CXXConfigure task has been modified to basically be a
wrapper around Mako, as it handles whether a file has changed or not.
As of right now, I'm unsure if I want to allow multiple inputs and
outputs, but that would rely on stuff not getting wonky, so I think a
one file per configure task focus would be great.