Removed the core.color module. Documentation will have to be updated
The color module now exists on PyPI, and can be used with the other
projects I now have. Additionally, the color module simply overrides the
print function, so it's usable elsewhere. This helps me out in that I
don't have to do selective 'print' calls, versus warning, error, etc.

I've updated the requirements, tox, and setup.* files to show this
change.

Additionally, I've decided that since pip is so important to what I
focus on and use, that I'll eventually be making it a dependency for
bit. Assuming someone was serious about python they would be using it
anyhow. The other tools I'm currently writing (CI server, etc.) are also
going to utilize the pip.util.find_command, otherwise I'd have to find a
place for my crappy and not all that impressive 'which' function. Oh
well!