Worked on MSVC and the CXX compile types.
As of right now, the CXXCompile task does nothing, but the compiler
classes are going to be where a majority of my attention is the next few
days. The compiler classes are where the information is 'transformed',
as well as performing basic environment setup (in the case of msvc,
where the environment variables are important).

As for the MSVC compiler, I'm not going to try to support VS2005. It's
2012, and if you're still stuck on it, then I'm sure you can find
another build system, or a better job. However, it probably wouldn't be
too hard to have a quick 'for-range' check instead of the hard-coded
variables, as the caching mechanism will only ever run once (hopefully).

Work will start on the linkers MUCH later, as I'm still thinking about
their interface.

I'm also a bit unsure as to how I'll be performing the actual execution
portion (I'd like to use concurrent.futures, but without passing a
class), as well as how I'll be getting the dependency information for
headers and whatnot. Hopefully, it'll be designed simply enough for me
to not worry about performance in that section, though I'm sure with a
large enough project things will start to bog down.